After WiFi.mode(WIFI_OFF) again re-init wifi mode
I encountered one problem where if i turn of the wifi hardware which is WiFi.mode(WIFI_OFF) or NULL command. And try to turn WiFi.mode(WIFI_STA) again it's not working.
Because of this line.
When you add it. You can turn on and turn off wifi from arduino libraries as many times as you would like.